Adding a confirmation prompt to delete the job

+++ src/views/JobView.vue
@@ -24,9 +24,12 @@ onMounted(async () => {
 
 const deleteJob = async () => {
   try {
-    await axios.delete(`/api/jobs/${jobId}`);
-    toast.success("Job Deleted Successfully");
-    router.push("/jobs");
+    const confirm = window.confirm("Are you sure you want to delete this job?");
+    if (confirm) {
+      await axios.delete(`/api/jobs/${jobId}`);
+      toast.success("Job Deleted Successfully");
+      router.push("/jobs");
+    }
   } catch (error) {
     console.log("error deleting job", error);
     toast.error("Job was not deleted");
